add the ability to skip vmware or virtualbox

+++ scripts/install-guest-additions.sh
@@ -12,21 +12,27 @@ box="$1"
 outbox="$2"
 vmware_box="$3"
 vmware_outbox="$4"
+skip_virtualbox="$5"
+skip_vmware="$6"
 
 name="$(basename "${outbox}" .box)-$$"
-
-# Ensure vbguest plugin installed.
-vagrant plugin list | grep vagrant-vbguest || vagrant plugin install vagrant-vbguest
-
-# Ensure vmware-desktop plugin installed.
-vagrant plugin list | grep vagrant-vmware-desktop || vagrant plugin install vagrant-vmware-desktop
 
 # Create temporary vagrant directory.
 export VAGRANT_CWD="$(mktemp -d -t "${name}")"
 
-# Register base box.
-vagrant box add --name "${name}" "${box}"
-vagrant box add --name "${name}" "${vmware_box}"
+if [ "$skip_virtualbox" = false ]; then
+  # Ensure vbguest plugin installed.
+  vagrant plugin list | grep vagrant-vbguest || vagrant plugin install vagrant-vbguest
+  # Register base box.
+  vagrant box add --name "${name}" "${box}"
+fi
+
+if [ "$skip_vmware" = false ]; then
+  # Ensure vmware-desktop plugin installed.
+  vagrant plugin list | grep vagrant-vmware-desktop || vagrant plugin install vagrant-vmware-desktop
+  # Register base box.
+  vagrant box add --name "${name}" "${vmware_box}"
+fi
 
 # Install security updates.
 # Install compiler and kernel headers required for building guest additions.
@@ -57,15 +63,23 @@ Vagrant.configure(2) do |config|
   end
 end
 EOF
-vagrant up --provider=vmware_desktop
-vagrant halt
 
-# Export box.
-vagrant package --output "${vmware_outbox}"
-vagrant destroy --force
+if [ "$skip_vmware" = false ]; then
+  vagrant up --provider=vmware_desktop
+  vagrant halt
 
-vagrant up --provider=virtualbox
-vagrant halt
+  # Export box.
+  vagrant package --output "${vmware_outbox}"
+  vagrant destroy --force
+
+  # Unregister base box.
+  vagrant box remove "${name}" --provider=vmware_fusion
+fi
+
+if [ "$skip_virtualbox" = false ]; then
+  vagrant up --provider=virtualbox
+  vagrant halt
+fi
 
 # Reboot in case of kernel security updates above.
 # Install guest additions.
@@ -83,18 +97,20 @@ Vagrant.configure(2) do |config|
   end
 end
 EOF
-vagrant up --provider=virtualbox
-vagrant halt
 
-# Export box.
-vagrant package --output "${outbox}"
+if [ "$skip_virtualbox" = false ]; then
+  vagrant up --provider=virtualbox
+  vagrant halt
 
-# Destroy VM.
-vagrant destroy --force
+  # Export box.
+  vagrant package --output "${outbox}"
 
-# Unregister base box.
-vagrant box remove "${name}" --provider=vmware_fusion
-vagrant box remove "${name}" --provider=virtualbox
+  # Destroy VM.
+  vagrant destroy --force
+
+  # Unregister base box.
+  vagrant box remove "${name}" --provider=virtualbox
+fi
 
 # Clean up temporary vagrant directory.
 rm -rf "${VAGRANT_CWD}"
